# src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/swagger.json": {
      "get": {
        "operationId": "SwaggerController_getSwaggerJson",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Swagger"
        ]
      }
    },
    "/user/profile": {
      "put": {
        "operationId": "UserController_changeProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "tags": [
          "user"
        ]
      },
      "get": {
        "operationId": "UserController_findOneProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "tags": [
          "user"
        ]
      }
    },
    "/user": {
      "get": {
        "operationId": "UserController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "tags": [
          "user"
        ]
      }
    },
    "/user/{id}": {
      "patch": {
        "operationId": "UserController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "tags": [
          "user"
        ]
      },
      "delete": {
        "operationId": "UserController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "tags": [
          "user"
        ]
      }
    },
    "/user/change-email": {
      "post": {
        "operationId": "UserController_changeEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/ChangeEmailDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeEmailDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "tags": [
          "user"
        ]
      }
    },
    "/user/verify-email-otp": {
      "post": {
        "operationId": "UserController_verifyEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/CheckOtpDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckOtpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "tags": [
          "user"
        ]
      }
    },
    "/user/change-phone": {
      "post": {
        "operationId": "UserController_changePhone",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/ChangePhoneDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePhoneDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "tags": [
          "user"
        ]
      }
    },
    "/user/verify-phone-otp": {
      "post": {
        "operationId": "UserController_verifyPhone",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/CheckOtpDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckOtpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "tags": [
          "user"
        ]
      }
    },
    "/user/change-username": {
      "post": {
        "operationId": "UserController_changeUsername",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/ChangeUsernameDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeUsernameDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "tags": [
          "user"
        ]
      }
    },
    "/auth/user-esxistence": {
      "post": {
        "operationId": "AuthController_userExistence",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/AuthDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/check-otp": {
      "post": {
        "operationId": "AuthController_checkOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/CheckOtpDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckOtpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/check-login": {
      "get": {
        "operationId": "AuthController_checkLogin",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/category": {
      "post": {
        "operationId": "CategoryController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Category"
        ]
      },
      "get": {
        "operationId": "CategoryController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Category"
        ]
      }
    },
    "/category/{id}": {
      "get": {
        "operationId": "CategoryController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Category"
        ]
      },
      "patch": {
        "operationId": "CategoryController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Category"
        ]
      },
      "delete": {
        "operationId": "CategoryController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Category"
        ]
      }
    }
  },
  "info": {
    "title": "Virgool API",
    "description": "NestJS Virgool API description",
    "version": "0.0.1",
    "contact": {}
  },
  "tags": [],
  "servers": [
    {
      "url": "http://localhost:3500"
    }
  ],
  "components": {
    "securitySchemes": {
      "Authorization": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token",
        "description": "Access token (Stored in cookie for secure access)"
      }
    },
    "schemas": {
      "ProfileDto": {
        "type": "object",
        "properties": {
          "nick_name": {
            "type": "string",
            "example": "john_doe"
          },
          "age": {
            "type": "number",
            "example": 25,
            "nullable": true
          },
          "bio": {
            "type": "string",
            "example": "Backend developer",
            "nullable": true
          },
          "image_profile": {
            "type": "string",
            "example": "https://example.com/profile.jpg",
            "nullable": true,
            "format": "binary"
          },
          "bg_image": {
            "type": "string",
            "example": "https://example.com/bg.jpg",
            "nullable": true
          },
          "birthday": {
            "format": "date-time",
            "type": "string",
            "example": "1998-05-12",
            "nullable": true,
            "description": "ISO date string"
          },
          "linkdin_profile": {
            "type": "string",
            "example": "https://linkedin.com/in/username",
            "nullable": true
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female",
              "other"
            ],
            "nullable": true
          }
        }
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {}
      },
      "ChangeEmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "test@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "CheckOtpDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          }
        },
        "required": [
          "code"
        ]
      },
      "ChangePhoneDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string"
          }
        },
        "required": [
          "phone"
        ]
      },
      "ChangeUsernameDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "DUMMY_USER"
          }
        },
        "required": [
          "username"
        ]
      },
      "AuthType": {
        "type": "string",
        "enum": [
          "login",
          "register"
        ]
      },
      "AuthMethod": {
        "type": "string",
        "enum": [
          "username",
          "phone",
          "email"
        ]
      },
      "AuthDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AuthType"
              }
            ]
          },
          "method": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AuthMethod"
              }
            ]
          }
        },
        "required": [
          "username",
          "type",
          "method"
        ]
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "priority": {
            "type": "number"
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "priority": {
            "type": "number"
          }
        }
      }
    }
  }
}